feat(generator): merge cssRegression generator into buildCSS generator
Improve the way the css regression scenarios are picked up.
Now tests are stores in
{moduleDir}/cssTest

feature/css-regression-tests

feature/css-regression-tests
