+++ lib/buildTool/webpack/buildCSS/buildCSS.js
@@ -15,6 +15,14 @@ module.exports = function() {
     let config = this.getGlobalConfig();
     let cssSourceFormat = config.buildCSS.sourceFormat;
 
+    if(config.buildCSS.useCssRegression) {
+      this.log('Writing CSS Regression tests configuration');
+
+      let toolResources = this.buildTool.getResources().cssRegression;
+
+      this.writeBuildToolConfig(toolResources);
+    }
+
     this.writeBuildToolConfig(toolResources.sourceFormat[cssSourceFormat]);
   }
