--- lib/buildTool/webpack/cssRegression/templates/exampleScenario.js
@@ -1,9 +1,0 @@
-module.exports = Object.assign({}, {
-  label: 'My Homepage',
-  url: '/',
-  selectors: [
-    'header',
-    'footer',
-  ]
-}, require('../scenarios.config'));
-
