--- lib/buildTool/webpack/cssRegression/templates/scenarios.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  readyEvent: null,
-  delay: 500,
-  misMatchThreshold: 0.1
-};
