--- lib/buildTool/webpack/cssRegression/templates/scenarios.js
@@ -1,3 +1,0 @@
-module.exports = [
-  require('./HomePage')
-];
