+++ lib/generators/buildCSS/index.js
@@ -28,6 +28,12 @@ module.exports = confitGen.create({
         message: 'CSS source format',
         choices: Object.keys(resources.sourceFormat),
         default: this.getConfig('sourceFormat') || resources.defaultSourceFormat
+      },
+      {
+        type: 'confirm',
+        name: 'useCssRegression',
+        message: 'Setup css regression tests?',
+        default: true
       }
     ];
