+++ lib/projectType/browser/latest.profile.json
@@ -9,7 +9,6 @@
     "buildCSS": "webpack",
     "buildHTML": "webpack",
     "buildJS": "webpack",
-    "cssRegression": "webpack",
     "entryPoint": "webpack",
     "paths": "",
     "release": "npm",
